Very hacky first shot at agent -> agent -> agent with auth.
The Birthday Agent can talk to a Calendar Agent for managing creating
calendar events. The Calendar Agent may need authorization to perform
its task, which the Birthday Agent needs to return to the end user
to perform.

I was unable to implement this flow "correctly" -- a correct flow would
have the Calendar Agent maintain its task response stream, but wait
asynchronously for authorization to come it. Once the authorization is
received, it could continue with its processing and resume publishing
task updates to its response stream. The Birthday Agent is unable to
handle this situation in the way its integrated with the Calendar Agent.
Using A2A as an ADK tool means that you have to produce a single
response for a task stream, which means you cannot return the
authorization URL while also waiting for the Calendar Agent to continue
its processing. So, instead the calendar agent breaks its response
stream once authorization is requested. Once authorization has been
performed, it's up to the user to prompt the Birthday Agent once more to
ask the Calendar Agent to try again.

There's a lot of very fragile glue code here. In particular, the A2A/ADK
layers (both client and server) are not robust at all. They break easily
and frequently.